--- DayEasy-Capstone/README.md
@@ -1,31 +1,0 @@
-# DayEasy-Capstone
-
-To get started:
-https://docs.expo.dev/get-started/installation/
-
-	1. Download Expo on command prompt:
-	"npm install --global expo-cli"
-
-	2. Download Yarn:
-	https://classic.yarnpkg.com/en/docs/install#windows-stable
-
-	3. Download React Navigation:
-	"yarn add react-navigation"
-	(Navigation allows the user to go back and forth between screens.)
-
-	4. Download Node.js LTS (Do not download "Current")
-	https://nodejs.org/en/download/
-
-Expo fundementals:
-	1. Uploaded PDF's on Teams
-	2. Templates on Teams
-
-Screen layout designs:
-Flexbox: https://reactnative.dev/docs/flexbox
-
-API Search:
-https://docs.expo.dev/versions/latest/sdk/webbrowser/
-
-
-Pedometer:
-https://docs.expo.dev/versions/v44.0.0/sdk/pedometer/
