--- DayEasy-Capstone/index.js
@@ -1,5 +1,0 @@
-import {AppRegistry} from 'react-native';
-import App from './App';
-import {name as appName} from './app.json';
-
-AppRegistry.registerComponent(appName, () => App);
