--- DayEasy-Capstone/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "dayeasy-fresh",
-  "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
-  "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "web": "expo start --web",
-    "eject": "expo eject"
-  },
-  "dependencies": {
-    "@react-navigation/native": "^6.0.8",
-    "@react-navigation/native-stack": "^6.5.0",
-    "expo": "~44.0.0",
-    "expo-application": "~4.0.1",
-    "expo-auth-session": "~3.5.0",
-    "expo-random": "~12.1.1",
-    "expo-status-bar": "~1.2.0",
-    "firebase": "^9.6.7",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "react-native": "0.64.3",
-    "react-native-safe-area-context": "3.3.2",
-    "react-native-screens": "~3.10.1",
-    "react-native-web": "0.17.1"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.12.9"
-  },
-  "private": true
-}
